Inline headers from physics
