+++ src/physics/em/AtomicRelaxation.hh
@@ -8,6 +8,7 @@
 #pragma once
 
 #include "base/Macros.hh"
+#include "base/MiniStack.hh"
 #include "base/Types.hh"
 #include "physics/base/CutoffView.hh"
 #include "physics/base/Units.hh"
@@ -78,6 +79,142 @@ class AtomicRelaxation
 };
 
 //---------------------------------------------------------------------------//
-} // namespace celeritas
+// INLINE DEFINITIONS
+//---------------------------------------------------------------------------//
+/*!
+ * Construct with shared and state data.
+ *
+ * The secondaries must have enough storage allocated for particles produced in
+ * atomic relaxation and the vacancies must have enough storage allocated for
+ * the stack of subshell IDs: this should be handled in code *before*
+ * construction.
+ *
+ * The precondition of the element having relaxation data is satisfied by the
+ * AtomicRelaxationHelper -- it is only "true" if a distribution can be
+ * emitted.
+ */
+CELER_FUNCTION
+AtomicRelaxation::AtomicRelaxation(const AtomicRelaxParamsRef& shared,
+                                   const CutoffView&           cutoffs,
+                                   ElementId                   el_id,
+                                   SubshellId                  shell_id,
+                                   Span<Secondary>             secondaries,
+                                   Span<SubshellId>            vacancies)
+    : shared_(shared)
+    , gamma_cutoff_(cutoffs.energy(shared_.gamma_id).value())
+    , electron_cutoff_(cutoffs.energy(shared_.electron_id).value())
+    , el_id_(el_id)
+    , shell_id_(shell_id)
+    , secondaries_(secondaries)
+    , vacancies_(vacancies)
+{
+    CELER_EXPECT(shared_ && el_id_ < shared_.elements.size()
+                 && shared_.elements[el_id_]);
+    CELER_EXPECT(shell_id);
+}
 
-#include "AtomicRelaxation.i.hh"
+//---------------------------------------------------------------------------//
+/*!
+ * Simulate atomic relaxation with an initial vacancy in the given shell ID
+ */
+template<class Engine>
+CELER_FUNCTION AtomicRelaxation::result_type
+AtomicRelaxation::operator()(Engine& rng)
+{
+    const AtomicRelaxElement& el     = shared_.elements[el_id_];
+    const auto&               shells = shared_.shells[el.shells];
+    MiniStack<SubshellId>     vacancies(vacancies_);
+
+    // Push the vacancy created by the primary process onto a stack.
+    vacancies.push(shell_id_);
+
+    // Total number of secondaries
+    size_type count      = 0;
+    real_type sum_energy = 0;
+
+    // Generate the shower of photons and electrons produced by radiative and
+    // non-radiative transitions
+    while (!vacancies.empty())
+    {
+        // Pop the vacancy off the stack and check if it has transition data
+        SubshellId vacancy_id = vacancies.pop();
+        if (vacancy_id.get() >= shells.size())
+            continue;
+
+        // Sample a transition (TODO: refactor to use Selector but with
+        // "remainder")
+        const AtomicRelaxSubshell& shell = shells[vacancy_id.get()];
+        const TransitionId trans_id      = this->sample_transition(shell, rng);
+
+        if (!trans_id)
+            continue;
+
+        // Push the new vacancies onto the stack and create the secondary
+        const auto& transition
+            = shared_.transitions[shell.transitions][trans_id.get()];
+        vacancies.push(transition.initial_shell);
+        if (transition.auger_shell)
+        {
+            vacancies.push(transition.auger_shell);
+
+            if (transition.energy >= electron_cutoff_)
+            {
+                // Sampled a non-radiative transition: create an Auger electron
+                CELER_ASSERT(count < secondaries_.size());
+                Secondary& secondary  = secondaries_[count++];
+                secondary.direction   = sample_direction_(rng);
+                secondary.energy      = MevEnergy{transition.energy};
+                secondary.particle_id = shared_.electron_id;
+
+                // Accumulate the energy carried away by secondaries
+                sum_energy += transition.energy;
+            }
+        }
+        else if (transition.energy >= gamma_cutoff_)
+        {
+            // Sampled a radiative transition: create a fluorescence photon
+            CELER_ASSERT(count < secondaries_.size());
+            Secondary& secondary  = secondaries_[count++];
+            secondary.direction   = sample_direction_(rng);
+            secondary.energy      = MevEnergy{transition.energy};
+            secondary.particle_id = shared_.gamma_id;
+
+            // Accumulate the energy carried away by secondaries
+            sum_energy += transition.energy;
+        }
+    }
+
+    result_type result;
+    result.count  = count;
+    result.energy = sum_energy;
+    return result;
+}
+
+//---------------------------------------------------------------------------//
+/*!
+ * Sample an atomic transition.
+ *
+ * TODO: refactor to use a Selector-like algorithm that allows a "remainder"
+ * that indicates "not sampled".
+ */
+template<class Engine>
+inline CELER_FUNCTION auto
+AtomicRelaxation::sample_transition(const AtomicRelaxSubshell& shell,
+                                    Engine& rng) -> TransitionId
+{
+    const auto& transitions = shared_.transitions[shell.transitions];
+
+    real_type accum = -generate_canonical(rng);
+    for (size_type i = 0; i < transitions.size(); ++i)
+    {
+        accum += transitions[i].probability;
+        if (accum > 0)
+            return TransitionId{i};
+    }
+
+    // No transition was sampled: skip to the next vacancy
+    return {};
+}
+
+//---------------------------------------------------------------------------//
+} // namespace celeritas
