+++ src/physics/em/detail/SeltzerBergerInteractor.hh
@@ -7,6 +7,8 @@
 //---------------------------------------------------------------------------//
 #pragma once
 
+#include "base/ArrayUtils.hh"
+#include "base/Constants.hh"
 #include "base/Macros.hh"
 #include "base/StackAllocator.hh"
 #include "physics/base/CutoffView.hh"
@@ -19,8 +21,12 @@
 #include "physics/material/MaterialView.hh"
 #include "physics/material/Types.hh"
 
+#include "PhysicsConstants.hh"
 #include "SeltzerBergerData.hh"
+#include "SBEnergyDistHelper.hh"
+#include "SBEnergyDistribution.hh"
 #include "SBEnergySampler.hh"
+#include "SBPositronXsCorrector.hh"
 #include "BremFinalStateHelper.hh"
 
 namespace celeritas
@@ -95,7 +101,80 @@ class SeltzerBergerInteractor
 };
 
 //---------------------------------------------------------------------------//
+// INLINE DEFINITIONS
+//---------------------------------------------------------------------------//
+/*!
+ * Construct with shared/device and state data.
+ *
+ * The incident particle must be within the model's valid energy range. this
+ * must be handled in code *before* the interactor is constructed.
+ */
+SeltzerBergerInteractor::SeltzerBergerInteractor(
+    const SeltzerBergerNativeRef& shared,
+    const ParticleTrackView&      particle,
+    const Real3&                  inc_direction,
+    const CutoffView&             cutoffs,
+    StackAllocator<Secondary>&    allocate,
+    const MaterialView&           material,
+    const ElementComponentId&     elcomp_id)
+    : shared_(shared)
+    , inc_energy_(particle.energy())
+    , inc_momentum_(particle.momentum())
+    , inc_direction_(inc_direction)
+    , inc_particle_is_electron_(particle.particle_id() == shared_.ids.electron)
+    , gamma_cutoff_(cutoffs.energy(shared.ids.gamma))
+    , allocate_(allocate)
+    , elcomp_id_(elcomp_id)
+    , sb_energy_sampler_(shared.differential_xs,
+                         particle,
+                         gamma_cutoff_,
+                         material,
+                         elcomp_id,
+                         shared.electron_mass,
+                         inc_particle_is_electron_)
+    , final_state_interaction_(inc_energy_,
+                               inc_direction_,
+                               inc_momentum_,
+                               shared.electron_mass,
+                               shared.ids.gamma)
+{
+    CELER_EXPECT(particle.particle_id() == shared_.ids.electron
+                 || particle.particle_id() == shared_.ids.positron);
+    CELER_EXPECT(gamma_cutoff_ > zero_quantity());
+}
+
+//---------------------------------------------------------------------------//
+/*!
+ * Bremsstrahlung using the Seltzer-Berger model.
+ *
+ * See section 10.2.1 of the Geant physics reference 10.6.
+ */
+template<class Engine>
+CELER_FUNCTION Interaction SeltzerBergerInteractor::operator()(Engine& rng)
+{
+    // Check if secondary can be produced. If not, this interaction cannot
+    // happen and the incident particle must undergo an energy loss process
+    // instead.
+    if (gamma_cutoff_ > inc_energy_)
+    {
+        return Interaction::from_unchanged(inc_energy_, inc_direction_);
+    }
+
+    // Allocate space for the brems photon
+    Secondary* secondaries = this->allocate_(1);
+    if (secondaries == nullptr)
+    {
+        // Failed to allocate space for the secondary
+        return Interaction::from_failure();
+    }
+
+    // Sample the bremsstrahlung photon energy
+    Energy gamma_energy = sb_energy_sampler_(rng);
+
+    // Update kinematics of the final state and return this interaction
+    return final_state_interaction_(rng, gamma_energy, secondaries);
+}
+
+//---------------------------------------------------------------------------//
 } // namespace detail
 } // namespace celeritas
-
-#include "SeltzerBergerInteractor.i.hh"
