+++ src/physics/material/MaterialTrackView.hh
@@ -7,8 +7,12 @@
 //---------------------------------------------------------------------------//
 #pragma once
 
+#include <cmath>
+
+#include "base/Assert.hh"
 #include "base/Macros.hh"
 #include "base/Types.hh"
+#include "physics/base/Units.hh"
 #include "MaterialData.hh"
 #include "MaterialView.hh"
 #include "Types.hh"
@@ -74,6 +78,71 @@ class MaterialTrackView
 };
 
 //---------------------------------------------------------------------------//
-} // namespace celeritas
+// INLINE DEFINITIONS
+//---------------------------------------------------------------------------//
+/*!
+ * Construct from dynamic and static particle properties.
+ */
+CELER_FUNCTION
+MaterialTrackView::MaterialTrackView(const MaterialParamsRef& params,
+                                     const MaterialStateRef&  states,
+                                     ThreadId                 tid)
+    : params_(params), states_(states), thread_(tid)
+{
+    CELER_EXPECT(tid < states.state.size());
+}
 
-#include "MaterialTrackView.i.hh"
+//---------------------------------------------------------------------------//
+/*!
+ * Initialize the particle.
+ */
+CELER_FUNCTION MaterialTrackView&
+MaterialTrackView::operator=(const Initializer_t& other)
+{
+    CELER_EXPECT(other.material_id < params_.materials.size());
+    this->state() = other;
+    return *this;
+}
+
+//---------------------------------------------------------------------------//
+/*!
+ * Current material identifier.
+ */
+CELER_FUNCTION MaterialId MaterialTrackView::material_id() const
+{
+    return this->state().material_id;
+}
+
+//---------------------------------------------------------------------------//
+/*!
+ * Get material properties for the current material.
+ */
+CELER_FUNCTION MaterialView MaterialTrackView::material_view() const
+{
+    return MaterialView(params_, this->material_id());
+}
+
+//---------------------------------------------------------------------------//
+/*!
+ * Access scratch space with at least one real per element component.
+ */
+CELER_FUNCTION Span<real_type> MaterialTrackView::element_scratch()
+{
+    auto            offset = thread_.get() * params_.max_element_components;
+    Span<real_type> all_scratch
+        = states_.element_scratch[AllItems<real_type, MemSpace::native>{}];
+    CELER_ENSURE(offset + params_.max_element_components <= all_scratch.size());
+    return all_scratch.subspan(offset, params_.max_element_components);
+}
+
+//---------------------------------------------------------------------------//
+/*!
+ * Access the thread-local state.
+ */
+CELER_FUNCTION MaterialTrackState& MaterialTrackView::state() const
+{
+    return states_.state[thread_];
+}
+
+//---------------------------------------------------------------------------//
+} // namespace celeritas
